+++ omicverse/utils/__init__.py
@@ -35,3 +35,4 @@ from ._knn import *
 from ._heatmap import *
 from ._roe import roe
 from ._paga import cal_paga,plot_paga
+from ._cluster import cluster